legal-ai-case-management-system/backend/requirements.txt
# For reading PDF files
PyPDF2

# For reading DOCX files
python-docx

# For AI and Natural Language Processing
transformers
torch
sentencepiece
tf-keras

# Note: scikit-learn is not used directly in the final script
# but can be useful for other NLP tasks. It can be omitted if not needed.